[feedback] Simplify AnnotationsProviders
After revisiting the annotation collection process I realized that it's
overly complex for what it needs to achieve. We were doing a ton of
unncessary and redundant filtering that made the process hard to
understand and reason about.

Instead of filtering the allowlist prior to instantiating each provider,
we just pass the allowlist to a provider it gets its annotations and
rely on the provider to do the filtering.

Additionally, we always instantiate all annotation providers instead of
instantiating the providers that have a non-empty intersection with the
allowlist. This reduces the complexity of annotation collection at the
cost of putting the burden of not collectiong annotations when the
intersection is empty on the providers

TESTED=`fx test feedback-data-tests`

Bug: 49922

Change-Id: I3a1c0731437c50f018e0ccd40e4b913d33d5b55a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382692
Commit-Queue: Alex Pankhurst <pankhurst@google.com>
Testability-Review: Alex Pankhurst <pankhurst@google.com>
Testability-Review: Francois Rousseau <frousseau@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>