[trace] support triggers (alert:action mapping)
This CL modifies the 'trace' util to implement triggers. A new
command line option is added to specify triggers, and the json 'spec'
allows triggers to specified.

TEST: fx test trace_tests
      Added unit tests for spec changes.
      Added a new integration test for triggers.

The new test provider runs for an hour, so that the alert is needed
to end the test. It also sets the session duration to an hour. A
trigger is specified to stop the session. Because triggers work, the
test terminates promptly. If triggers didn't work, the test would run
for one hour, which would cause the test runner to time out and
declare the test failed.

Bug: 49629

Change-Id: I0777e02dde3b773f50574358e35b2242f3f9869e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382766
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>