[camera][controller] Start/stop streams irrespective of their current state
The semantics of Start/StopStream()
include the legality of starting/stopping a stream which has already
started/stopped.

This CL adds the ability to resume streams which clients created and
started while the device was in a muted state.

Prior to this change, the camera subsystem incorrectly handled the muted
device state. In this state, the sensor did not start streaming when
clients requested streams to start. However, the controller's state
indicated that these streams had, in fact, started.

Later, when muting stopped, all existing streams should have started to
stream. The controller did not forward these requests to the ISP driver,
as it believed the streams had already started.

This CL fixes that behavior and ensures that the controller will request
the ISP driver to start/stop the current streams upon unmuting.

Fixed: 50403

Test: fx test camera_tests
      Started camera-gym app while device was muted. Streaming
      resumed when device was unmuted.

Change-Id: Ie46d599267fd28dbd3092c675e5b853e988883a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382328
Reviewed-by: Brijen Raval <braval@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>
Testability-Review: Brijen Raval <braval@google.com>
Commit-Queue: Brijen Raval <braval@google.com>