[cobalt] Changes variable name to |validated_clock|.
This came up while attempting to understand some code.
Within cobalt_app.cc there is an instance of FuchsiaSystemClock
which implements the interface ValidatedClock. A pointer to this
is passed to CreateCobaltConfig where it is used as the value
of the |validated_clock| field. There are also instances of
std::chrono::system_clock and util::SystemClock used in the
same class. There was (prior to my change) a variable called
system_clock of type util::SystemClock even though there was
also a parameter called system_clock of type FuchsiaSystemClock.

I found all of this difficult to follow. I think it is a little
easier to read if the instances of FuchsiaSystemClock are called
|validated_clock|.

I also changed a log statement in the test_app.

Change-Id: Ib8fc4cd470ecb4ede9ea92685cc4c2081e2fee49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383335
Reviewed-by: Cameron Dale <camrdale@google.com>
Testability-Review: Cameron Dale <camrdale@google.com>
Commit-Queue: Mitch Rudominer <rudominer@google.com>