[intl] Skeleton implementation of the l10n lookup
This is the first part of the implementation of the l10n lookup for the
screen reader.  For the time being it contains a dummy
implementation, which is to be expanded in the followup work, and a fake
implementation which will be used in tests.

Bug: 50725
Change-Id: I7b4db3e3b4d3c908f888459e37404937abe6e77a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383088
Testability-Review: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>