[modular][sdk] Mark ComponentContext.ConnectToAgent() transitional.
There is an out-of-tree client that defines a Fake version of this
protocol. In order to remove the methods in the .fidl, we must allow
that client to build without defining ConnectToAgent().

Bug: 49976
Change-Id: I08a5daf79b72f031636ff2bd1d410b5f2ca6877f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383577
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
API-Review: Kevin Lindkvist <lindkvist@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Gabe Schine <thatguy@google.com>