[runtests] Avoid infinite loop
When an empty VMO is published, we would continue the loop, but
since we only popped the element from the list at the end of the
loop, this would never happen and we would enter an infinite loop.
So instead, pop the element at the beginning of the loop. We also
deduplicate the VMO mapping logic between the initialization and
merge loop.

Change-Id: I84c48d039821bdc5840694949a57b300a37bd8de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383543
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Petr Hosek <phosek@google.com>