[blobfs][rac] Test `ZSTDRead()`
This change contains two tests and a bug fix. The bug is an incorrect
calculation of how many blocks to read in `ZSTDRead()`. This calculation
is currently based on the size of the read. It should be based on the
block-aligned offsets in the blob of the start and end of the read.

For testability, `ZSTDSeekableFile`, `ZSTDSeek()`, and `ZSTDRead()` are
now public, rather than hidden in `zstd-seekable-blob.cc`.

BUG=50837
BUG=44820
TEST=`fx shell runtests -t blobfs-unit-test`

Change-Id: Iff71e0fc01f84bfd120732405b94708e617cf35b