[tools][debug][elflib] Handle PE-COFF PDB GUIDs as build IDs
This makes elflib a bit of a misnomer, since it reads both ELF
and PE-COFF now.  It's really buildidlib anyway, more of its code
is about build IDs abstractly than about the ELF format.

PE-COFF debug files (PDB files) are in a disparate format that is
complicated to decode and doesn't have a handy Go library.  So
instead of decoding PDB files to match the build ID (GUID + age) in
the PE-COFF binary, a simple "is a PDB" detector is provided so that
buildidtool can do a kludge to suppress the build ID matching and
still link the .pdb file as the debug file.

Change-Id: I9589853640cd56c0ffc116d6340606eca13031a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381556
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>