[bt][hci] Avoid assuming kernel message batching behaviour in tests.
The current test "HCI_CommandChannelTest.EventHandlerResults" implicitly
assumes a particular behaviour of how the Zircon kernel will notify
userspace about batches of received messages in a channel. This
behaviour is changing, causing the test to start failing.

This CL moves the event callbacks in the test to use their own
loop/dispatcher, so that the test can independently control when the
event callbacks are processed (as is required for the test) instead of
this implicitly occurring at the end of each batch of messages on the
channel.

We additionally split out testing of "SendCommand" behaviour when an
event handler has already been registered into its own test.

Bug: 50601

Change-Id: Ic270b4dedac9cd53c75ff231b938031b2d233cb8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381895
Commit-Queue: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>