[storage] Add Zero method to storage::BlockBuffer.
This is needed for following changes. Where the buffer is backed by a
VMO we can use a specialised zero method. Also change the interface so
that it's block based like the rest of the methods.

Also renamed VnodeBufferType -> ResizeableBufferType since it's not
specific to a Vnode.

Removed the unit tests for VnodeBufferType since we should alrady have
coverage for the Zero method elsewhere.

Bug: 42096
Change-Id: I567087fc1bd3880b6ca980689d6460e0027e4d67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382651
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>