[fidlc] Recoverable parsing at member boundary
This CL builds on the functionality added in fxrev.dev/378714 to add
recoverable parsing between members of a declaration. E.g., if a table
member is missing an ordinal, the parser will attempt to recovr to the
next member of the table and continue parsing, rather than skipping to
the end of the table.

In future CLs:
* Recovery within a parameter/attribute list
* Token insertion
* Recoverable error reporting in the compilation stage

Test: fidlc host tests
Change-Id: I54916ab7f78766519371a39ec3891d9c781eab07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381145
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>