[images] support firmware_prebuilts
Adds support for the firmware_prebuilts variable, which allows adding a
set of firmware images to the build package.

This is a replacement for the existing bootloader_prebuilt variable
with the added ability to specify multiple firmware images of different
types - for example, for multi-stage bootloaders provided in separate
files.

Bug: 45606
Change-Id: I86cb7f8e49ce1ee853ff49e71af638140084aa9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381817
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: David Pursell <dpursell@google.com>