[fidlcat] use printer everywhere
This is a refactoring to be more consistent and versatile.

The printer is now independent of the decoder. That means that we can
use a printer when the decoder is deleted and we have just kept an
event. This will be used to create graphs.

The behavior is unchanged.

Change-Id: I69c6648b28ff5fa485781fdc06b487bba3eb7d1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383798
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Vincent Belliard <vbelliard@google.com>