[archivist] Don't fail on send if receiver is closed
The new API might abort the archivist future while the handling of
an event is going on on a separate task. When aborted the receiver end
of the channel is closed, this leads to SendError (can only occur when
the other end of the channel is missing). This caused a flake in logs.

Bug: 50850
Tested: existing tests

Change-Id: I740a9a2079f2854c8ac51305887c0124ccdd7a8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383875
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Luke Nicholson <lukenicholson@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>