[cobalt] Make FuchsiaSystemClock thread safe.
A pointer to an instance of FuchsiaSystemClock is passed in
to the CobaltConfig that is passed to the CobaltService
in Cobalt core. Currently the now() method is only ever
invoked from the Logger::Log*() method and so it is only
ever invoked on the main thread of the Cobalt FIDL service.
However in the future I can imagine we might want to invoke
the now() method from one of Cobalt core's background threads.
This CL makes it safe to do so.

Change-Id: I228320fad3837cf107a5cb946b6bbf6a43f62cf8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383354
Reviewed-by: Cameron Dale <camrdale@google.com>
Testability-Review: Cameron Dale <camrdale@google.com>
Commit-Queue: Mitch Rudominer <rudominer@google.com>