[media][audio_core] Parse config driver_gain_db
Parse a new driver_gain_db field from the output device configuration
and include it in the device profile. In a following CL, the driver gain
value will be applied to the driver upon initialization.

This is the first in a chain of CLs that will allow driver gain to be
modified in AudioCore on a per-device basis and in relation to audio
effects. Enabling driver_gain setting in the device configuration
ensures the driver's default gain setting will be Unity (0 db) unless
modified in the configuration.

Driver gain modifications are meant to be temporary fixes for loudness
issues and/or to compensate for a lack of device tuning.

Bug: 49302

Test: fx test audio_core_unittests
Change-Id: I074f099c0d12c1f7ce27ece45dcb24c691ee5825
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383403
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Payton Bradsky <pbradsky@google.com>