[fidlgen][libfuzzer] Generate goldens for libfuzzer
This CL fixes a bug in regen.sh which prevented it from generating
libfuzzer goldens. It is only supposed to emit them for FIDL libaries
that have a nonempty protocol; the bug was that it tested for this
assuming $src_path was a FIDL file, when it is actually JSON IR.

This CL also fixes some inaccuracies in the contributing_to_fidl.md
goldens table.

Change-Id: Ia6a56be67c00918c9af7d94c9f0dcf5f035c2cec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383906
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>