[build][unification] Make kazoo available to GN build.
The kazoo tool is used to generate Zircon syscall bindings.
It is used by the ZN build extensively, but this CL makes it
also available to the GN build.

Note that this does not migrate the tool from the ZN build
to the GN build. Instead it creates a new GN host executable
(out/default/host_x64/kazoo) that duplicates the functionality
of the ZN one (out/default.zircon/tools/kazoo).

This will be used in a future CL to generate Rust bindings
directly in the GN build.

- Add missing unification wrappers for zx_host_tool()
  and zx_test() under //build/unification/

- Add zx_host_tool_action_foreach() template to wrap the ZN
  host_tool_action_foreach() and provide a GN implementation
  for it based on compiled_action().

- GN now supports dependencies to out-of-tree targets, so
  remove the rapidjson defines/include_dir hack in
  zircon/tools/kazoo/BUILD.gn and just depend on
  "$zx/../third_party/rapidjson" instead.

- Ensure the kazoo library links with the appropriate cmdline
  target, depending on the build system being used.

- Use different names for the generated ZN and GN kazoo host
  binaries to avoid errors when running CQ infra scripts which
  expect only a single instance of each host tool name.

Bug: 43917
Testing: fx test //zircon/tools/kazoo:tests

Change-Id: Ifda79d17d3e7281804262119803698153cef9242
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383433
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>