[syscalls][kazoo] Fix a few Go remapping errors
There were a few types that didn't have appropriate name hacks added for
them. Kazoo errors out on attempting to run regeneration, but because Go
isn't run as part of the build (it's in a separate repo) they weren't
noticed.

As a way of avoiding this in future, generate all possible output styles
to a golden file, and diff in a build rule to catch failures in all
output styles, as well as potentially unintentional changes.

This has the possibly unfortunate side-effect of catching simple
intentional changes when, e.g. a syscalls arguments change, but it's
probably a reasonable idea to have an second intentional review of the
output generated for those changes anyway.

Test: zircon/tools/kazoo/golden-test added
Change-Id: Ic32dbd9d333cf182e20c34c0559a8f9fc3d37386
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379092
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Testability-Review: Scott Graham <scottmg@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>