[appmgr][test] Fix flake in component events integration test
This test was completely wrong. It was connecting to
Componenteventprovider on an environment but not coming from a
component. Hence the connection was failing as the component
NO_COMPONENT is not allow-listed to connect to ComponentEventProvider.
This was the cause of the flake. As a fix, the test was replicated but
using a real component.

Tested: component_events_integration_test
Fixed: 49600

MULTIPLY: component_events_integration_tests: 100


Change-Id: I16294b06436e5181690c113f2fb0142072989d83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384136
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>