[media][audio_core] Enable driver_gain_db on init
On driver initialization, driver_gain_db is applied to the driver. The
driver_gain_db value is sourced from the device config, and if the key
is missing in the config or the config is not available, a default gain
value of 0.0db is applied.

Bug: 49302

Test: fx test audio_device_tests
Change-Id: I7ee78eb7a60e8aa3cb805ab76840f17c7499f2f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383622
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Payton Bradsky <pbradsky@google.com>