[ffx] Add simple e2e test.
This is the simplest possible useful E2E test, but we'll extend this
with more soon.

Testing note: we can't use `MULTIPLY` here, as the emulator must be
rebooted in between test runs for the test to pass.

Change-Id: If16d0ff776ae912726540dceba15f938b53fdedd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381217
Commit-Queue: Jordon Wing <jwing@google.com>
Testability-Review: Andrew Davies <awdavies@google.com>
Reviewed-by: Adam Barth <abarth@google.com>