[fx][test] Updating klog test for new host tool location
The location of loglistener changed from zircon to an installed host
tool. This updates the mock to reflect the new location.

Testability: This is a test that was failing before updating the new
location, now it passes.

Change-Id: I1a6640064679108f21e2e421178387d602415c18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384353
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>