[kernel][x86] Use para-virtualized IPI under KVM
When available, use para-virtualized IPIs. This allows a hypervisor to
optimise the IPI path, as it is directly provided a CPU mask of all the
vCPUs to send IPIs to through the the use of a VMCALL instruction.

Bug: 49133
Change-Id: Ibc14021b27887e9bd2f7f2ce9e8c8bdb5b9e5240
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383679
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>