[x86][pci] Introduce PCI RootHost
PciRootHost replaces the static allocators used in the
x86 board file with an interface that can be implemented
on different platforms for PCI Root support. Additionally,
Pciroot is now templated on the type of platform context necessary,
so bringing up PCI on a non-ACPI system should be similar to setting
up the basic board file as one would for a normal driver. For any
architecture to support PCI it will now only need to provide a
Pciroot implementation for that context and instantiate a RootHost.

Some other changes were made to make the resource logic in the ACPI
parsing more clear to folks who don't speak ACPICA.

Change-Id: I5994b056e39f478653d09cef1f589c0f80bcc206
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/333310
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>