[fbl][containers] Change copy/move behavior of container nodes.
Change the default copy/move behavior of intrusive container nodes in
fbl.

:: Background ::
Nodes which exist in intrusive containers cannot be simply copied or
moved without introducing logically inconsistent states.  If a node,
A, from a doubly linked list were to be used to copy construct a node
B, we would have a situation where B thought it was in a list in A's
place, but the nodes that B was linked to would be linked back to A,
not B.  Even worse, if both A and B were in containers and A's node
were to be copy assigned into B, we would have a situation where the
nodes around B pointed to B, but B pointed to the nodes around A
instead.

Simply put, it makes no sense to allow copying from or assigning to
nodes which currently exist in containers.  Likewise, there is no good
way to define "moving" that makes sense in these situations either.

That said, it should be OK to copy or move a T which contains a node,
as long as the source and destinations are not currently in
containers.

Nick addressed the immediate problem in
I3a5eca29843767b9f263d982779f44233c4d8250 by changing the copy/move
construction behavior of our container nodes.  Copy/move constructed
nodes are now always default initialized (eg; not in a container)
while copy/move assigned nodes always hold their existing state.  In
addition, if the source of construction, or either the source or
destination of assignment are currently in containers, a DEBUG_ASSERT
will fire.  This allows Ts to be copied or moved, just not while
in a container (lest the ASSERTs fire).  In builds with ASSERTs turned
off, the copies can take place, but at least the integrity of the
containers will be preserved.

As it turns out, however, this behavior in our existing codebase is
actually very uncommon.  The places where it was happening tended to
be simple mistakes, not actually users who wanted to copy the contents
of a T while it was currently in a container.  For example, the
following code accidentally produces a copy of each member of the list
being enumerated:

  for (auto X : list_of_xs_) { ... }

It is almost certainly the case that what the user _meant_ to say was

  for (auto& X : list_of_xs_) { ... }

:: The Change ::

So, this CL introduces some new flags and changes the default behavior
of intrusive container nodes in fbl in order to help find mistakes
while still allowing flexibility in behavior.

Nodes are no longer copy/move construct-able/assign-able anymore.
Attempting to expand any of these operations will (by default) trigger
a static assert telling the user that this operation is not allowed.

In situations where the user _does_ want this behavior, there are now
4 new NodeOption flags which can be used in order to opt-in to
disabling the safeties.  They are

++ fbl::NodeOptions::AllowCopy
++ fbl::NodeOptions::AllowCopyFromContainer
++ fbl::NodeOptions::AllowMove
++ fbl::NodeOptions::AllowMoveFromContainer

Passing Allow(Copy|Move) will disable the static assert for the
associated constructor/assignment operations, but will retain the
ASSERTs.  Effectively, this allows the user to make copies or move
their Ts, but only when the instances in question are not current in
containers (enforced using DEBUG_ASSERT).

Going further, the user can use the "FromContainer" version of these
flags to disable the ASSERTs as well.  This allows a user to copy/move
from or even assign to their Ts while they are still in a container.
The behavior of the operations remains unchanged, however.  Assigning
to a T which is in one or more containers will do nothing to T's node
fields.  Neither the source nor the destination will be put into or
taken out of any containers as a result of assignment.  Similarly,
copy/move construction will never put the constructed T into any
containers, not will it change anything about the source T when moved.

:: Testing ::

Because of the sheer number of template options which need to be
tested here, the existing NodeState tests which were previously a
member of the base test environment container have been replaced with
a set of unit tests as a part of intrusive_container_node_tests.cc.
This makes it just a bit easier to manually exercise the many
different negative compilation tests which need to be exercised when
dealing with these options.

Tests: New unit tests.
Bug: 49643
Change-Id: I0b3f15f2fcd178e114f9090b35aabcf54fb86476
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381337
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>