[fidl][measure-tape] Generalize expression code
Instead of generating C++ expressions, we represent expressions in an
AST and later transform these into valid C++. For instance, the
expression

    FIDL_ALIGN(value.length() * 16)

is represented as the AST

    FidlAlign
    - Mult
      - Length
        - Local of "value"
      - Num of 16

This paves the way to a direct translation to Rust, while reusing the
biggest part of the code which emits the expressions and code blocks.

This representation also opens the door to simple constant folding, and
combining statments.

Change-Id: I870f0e8c8e39cb999ad1324d87a235d1cdbd07b7
Test: fx test measure-tape-tests
Bug: 50011
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381753
Commit-Queue: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>