[logging] Make fxl_logging a source_set
fxl_logging has always been a source_set on host. However,
on Fuchsia was defined as a shared_library because it contains
global state. With the recent migration of the FXL library to
use syslog, we can just rely on the global state in the syslog
shared library and not keep any state in FXL.

This CL introduces separate logging backends for host and
Fuchsia to reduce the number of ifdefs in the code. There is
opportunity to encapsulate more functionality in the backend
interface but this CL only introduces functions related to
modifying the global settings.

fxl::GetLogSettings() has been removed because we can't get the
log file path from the syslog library. The few clients that
used this function have been converted to not depend on it.

Bug: 40982
Change-Id: I6dd1c5ba3b5d9034adcd1c09430f11c85962f39e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384221
Commit-Queue: Saman Sami <samans@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Testability-Review: Adam Barth <abarth@google.com>