[brcmfmac] Fix alignment for sim IovarSet()
The sim implmenetation of IovarSet() relies on being able to cast
directly into the buffer provided, which may be misaligned for the
desired access.  Fix this by making a suitably aligned copy of the
buffer before passing it into IovarSet().  While this is another copy
step, it should not be significant for a testing framework such as sim.

Bug: 45759
Test: fx run-test brcmfmac_sim_test
Change-Id: I10f6fb5243b37e69f73b69e8b2f0904e93fcd65a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383904
Commit-Queue: John Sheu <sheu@google.com>
Reviewed-by: Joshua Conner <joshconner@google.com>
Testability-Review: Joshua Conner <joshconner@google.com>