[docs] Rewrite and improve sysmgr-configuration.md
This updates sysmgr's primary documentation, adding a lot of missing
detail about the configuration format and in general document more of
the pitfalls to avoid when modifying sysmgr configuration. A new section
was also added to document how to add to sysmgr's configuration and the
recommended best practices for avoid sysmgr config conflicts.

Bug: 48223
Bug: 48215
Change-Id: I128a4bc2188e968a54082d28b9817240e6b00ce0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374081
Commit-Queue: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: Shai Barack <shayba@google.com>