[mediaplayer] fix test flake
mediaplayer end-to-end tests are necessarily sensitive to real time,
because we don't have a distributed fake clock. The test bot
environment is such that the fake audio renderer can get behind to
the point that the VMO used for audio packets fills up and packets
are dropped. Verifying the contents of the packets is therefore
flakey.

This CL removes the check that verifies that actual audio packets in
the offending PlayBear test.

TEST: Change to test

Fixed: 50016
Change-Id: I3116b6ce420843eea6b7dde64e24a57a1292197a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384684
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>