[blobfs] Increase chunked compression level.
Value increased from 3->14, which is the smallest value that does not
increase the content or image size for blobfs per the baseline.

This will result in slowdowns on the blob installation path --
compression will be an order of magnitude slower. However, note that
installation is still IO bound even after this change.

Change-Id: Ied2dcfa1388ea0cfae8847baaee21d7c83a095fd