[fx] fix fx status when fx is not in path
`fx status` calls `fx gen` internally, but if `fx` is not in the PATH,
it fails with an exception.

This CL changes fx status to look for fx in
$FUCHSIA_DIR/.jiri_root/bin/fx instead, which is consistent with
`fx test`.

In the future, `fx` reentry and other util functions should be moved
to a common reusable library.

Test: remove .jiri_root/bin from PATH and run
  `.jiri_root/bin/fx status`. Before this CL it crashes with the
  message in fxb/51110. After this CL it works as expected.

Fixed: 51110
Change-Id: Ia4cdbca71e07e0171f0a339d9768b40a052ecbb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384761
Commit-Queue: Renato Mangini Dias <mangini@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Craig Labenz <craiglabenz@google.com>