[utest][fs] Remove race in mmap_crash tests
After performing munmap there is no guarantee that any operation (from
any thread) that performs a mapping or creates a VMAR will not reuse
the portion that was freed. As ASSERT_DEATH creates a thread, which
requires creating several memory mappings, it is possible for the
mapping to be reused between the munmap and the access.

This change moves the munmap into the death callback to remove any
mapping calls from happening before the access test can happen.

Bug: 41601

Change-Id: Id9ba4763efe4b501454810aeb1e98b7b8bf5f283
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384774
Testability-Review: Adrian Danis <adanis@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>