[kernel][heap] Update cmpct_get_info()
This change
(1) renames "size_bytes" to the less ambiguous "used_bytes";
heap_get_info()'s size_bytes also is updated to output the "total bytes"
(which "used" + "cached", which is a more informative quantity at that
level.
(2) extends the signature to also fill in the value for the size of the
cached allocation
(3) permits null parameters (which are ignored) for convenience.

Bug: 49123
Change-Id: I1b72a09ecf1130299ff7bd1c7c1f8f1db96362ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383597
Testability-Review: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>