[fbl][wavl] Remove a useless and incorrect using directive.
There is no real reason for the generic node_state fetcher in WAVL
tree to have a RankType using alias.  In addition, this alias was
incorrect.  It was missing the TagType parameter for the node_state
template which caused it to incorrectly fall back on the default tag.
This would only be caught when attempting to instantiate a tree of
objects which used the ContainableBaseClasses pattern, but no default
WAVLTreeContainable mix-ins.

Also, add new tests which would have caught this in the first place,
and fix an existing test which was broken and caught by the new test
augmentations.

Test: new tests
Change-Id: If7a83d749feeaf16588ad85fc3ad2fb767b95b9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383486
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>