[netemul] Moves gRPC client stress test to flaky.
The stress test repeatedly sends small messages to a guest
VM.  This test was written to demonstrate testability of a
change that fixed a race condition in this communication
mechanism.  In order to attempt to excite the issue, 1000
iterations of the gRPC operation are performed serially.
This test was written before the 5 minute timeout was
imposed on individual test runtime.  It seems that the test
has been exceeding the timeout a couple times per week.
This change moves the sometimes long-running test to flaky.

Bug: 49752
Change-Id: I46cdf1123cf0529604166c84502f2f265c4aab1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384637
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>