[swd-e2e-tests] Delay closing sl4f connection
This patch helps to work around http://fxbug.dev/47145, which is a
complicated series of bugs which resulted in the e2e test to fail with
random ssh client-side or server-side disconnections. The root cause of
the problem is how the e2e tests launched sl4f. We launch sl4f using
`run fuchsia-pkg://fuchsia.com/sl4f/0#meta/sl4f.cmx`, which runs sl4f in
the foreground. When we closed the session, it left the underlying ssh
channel in openssh in a weird state. Left in that state it's possible
for this to result in the client to err out, or server-sent keepalives
to be dropped.

Ideally we'd instead switch to daemonizing sl4f, but implementing that
is blocked by http://fxbug.dev/51195.

This patch instead works around that stack of bugs by just not closing
the ssh session until we detect a disconnection. This leaves the ssh
channel in a healthy state, which shouldn't cause the disconnections.
This can be undone once 51195 or any of the other 47145 bugs are fixed
fixed.

Test: This was extensively tested with a highly instrumented fuchsia,
go-ssh, and openssh-portable. To reproduce the failure before this patch
in a relatively consistent manner, we need to run `fx qemu -N` in one
window, and this in another.

```
% $(fx get-build-dir)/host_x64/system_upgrade_test \
  --device step-atom-yard-juicy \
  --fuchsia-build-dir $(fx get-build-dir) \
  --cycle-count 10000 \
  --cycle-timeout 10h
```

This should result in the test failing at random points in time. The bug
is more likely to occur if a 20 second sleep is added to the
implementation of /pkgfs/ctx/validate/missing, and for
tools/net/sshutil/Client.go's keepalive to be sent every 30 seconds.
This test should not fail with this patch applied.

Bug: 47145
Bug: 51195
Fixed: 51197

Change-Id: I467a550bafcb2e89eae378b4d396a84ed8a99c62
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384994
Reviewed-by: Kevin Wells <kevinwells@google.com>
Testability-Review: Kevin Wells <kevinwells@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>