[kernel] Remove unused "target_set_debug_led" method.
"target_set_debug_led" allows targets to implement methods for switching
a debug LED on and off. Currently, we only call it in the scheduler (to
indicate if the system is idle), but don't actually implement it for any
target or platform.

Remove it for now, and add it back later if the need arises.

Change-Id: I628c97251212ea048c787ba6616c4e500d598bc2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383973
Commit-Queue: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>