[kernel] Adjust kernel init levels.
In particular, add a "{arch,platform}_prevm_init" functions that are
called before virtual memory is (fully) enabled, but after the command
line and UART is available.

This is to facilitate a non-trivial chunk of code that runs on x86 which
needs access to the command line but also needs to run prior to
code-patching.

Change-Id: I64e79e2642640bba696804a22ddee64e9d221bce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383692
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>