[isolated-ota] launch the package resolver
This CL adds support for launching the package resolver
and package cache inside a sandboxed environment, against
the pkgfs already launched and using a provided repo configuration.

We use the integration test variants of pkg-cache and pkg-resolver so
that we can override paths like /config/data with the configuration
that's been provided. To ensure that these are available, we add an
'isolated-swd' package that the 'isolated-ota' library depends on that
contains all the relevant binaries and manifest files.

Change-Id: Ifc17e06262ab5712ec36d3ddb9d13b66c3f0f1f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377857
Testability-Review: Simon Shields <simonshields@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>